Update LAA test COE files
laa_modified_asm.coe should be good for a full
LAA test once we have UART ready.

laa.coe is a super simplified version of the LAA
test that doesn't use any predefined arrays but
instead just loads the registers directly. It's
good for testing that the register writing for
the main core even works.