Finalize LAA COE files for simulations
laa_manual.coe is the file used for simulations
in the presentation.

I ended up just writing the full COE file from
scratch to make sure none of the registers were
written from the RISC-V core to the LAA before
they had been committed. That's probably a bug
that should be looked into eventually, but it
works for now.

Also fixed a small bug in the LAA_core that Manu
found that prevented the reading of LAA registers
into the RISC-V core registers.